requirements.txt
dlib
numpy
opencv-contrib-python
Pillow
scipy
torch
torchvision
